enemies colliding with enemis doesnt hit our patch, in theoryu shoudl reove voice chat ui of dead players on reset, might have fixed the bug where sometiems on next level your camarea gets stuck in spectating mode

+++ OPJosMod/Patches/EnemyAIPatch.cs
@@ -36,7 +36,7 @@ namespace OPJosMod.GhostMode.Patches
         {
             mls.LogMessage("enemy collide with player patch hit");
 
-            if (PlayerControllerBPatch.isGhostMode)
+            if (PlayerControllerBPatch.isGhostMode && __instance.GetClosestPlayer().playerClientId == StartOfRound.Instance.localPlayerController.playerClientId)
             {
                 throw new Exception("player is a ghost don't try to kill again");
             }
